color HiPS fix

+++ src/render/src/shaders/hips/rasterizer/grayscale_to_colormap.frag
@@ -1,8 +1,8 @@
 #version 300 es
-precision highp float;
+precision mediump float;
 precision lowp sampler2D;
 precision lowp isampler2D;
-precision highp int;
+precision mediump int;
 
 in vec3 frag_uv_start;
 in vec3 frag_uv_end;
